requirements.txt
gradio
fal_client
requests
Pillow